worked on codesmells, moved tests into our component folder
