--- src/__test__/components/AssociateCard/AssociateCardFactory.test.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import {shallow} from 'enzyme';
-import {AssociateCardFactory} from '../../../components/AssociateCard/AssociateCardFactory'
-
-let wrapper:any;
-
-beforeEach(() => {
-    wrapper = shallow(<AssociateCardFactory />);
-});
-
-test("content is empty without any batch information given", () => {
-    const contentText = wrapper.find("#contentHolder").render().text();
-    expect(contentText).toBe("");
-});
